% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barGraph.R
\name{barGraph}
\alias{barGraph}
\title{Produce a bar graph}
\usage{
barGraph(x, y, mainTitle = "", xLabel = "", yLabel = "", yType = "int",
  yDigits = 0, barColor = "#009DDC", lineColor = barColor,
  lineWidth = 0, orderBy = NULL, orderDesc = FALSE, horizontal = FALSE)
}
\arguments{
\item{x}{character, factor, numeric, logical or Date vector}

\item{y}{numeric vector}

\item{mainTitle}{character string; graph title}

\item{xLabel}{character string; x-axis label}

\item{yLabel}{character string; y-axis label}

\item{yType}{character string; y-axis format type; valid 
values are \code{"int"}, \code{"dlr"} and \code{"pct"}}

\item{yDigits}{non-negative integer value indicating the number of decimal 
places to show when hovering over the bars}

\item{barColor}{character string; fill color for bars (valid color)}

\item{lineColor}{character string; line color (valid color)}

\item{lineWidth}{non-negative integer value indicating line width (use 0 to omit)}

\item{orderBy}{a value of \code{NULL} (do not order), \code{"x"} (order by x) or 
\code{"y"} (order by y)}

\item{orderDesc}{logical value; when ordering by \code{x} or \code{y}, order descending?}

\item{horizontal}{logical value; create a horizontal bar graph?}
}
\value{
A \code{\link{plot_ly}} bar graph using \code{x} for the bar labels 
and \code{y} for the heights of the bars
}
\description{
\code{barGraph} produces a simple bar graph
}
\examples{
barGraph(c("giraffes", "orangutans", "monkeys"), c(20, 14, 23))
}
\seealso{
\code{\link{plot_ly}}
}
