% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_pt.R
\name{boxplot_pt}
\alias{boxplot_pt}
\title{Boxplots With Data Points}
\usage{
boxplot_pt(
  y,
  x,
  jitter_amount = NULL,
  bp_col = 0,
  bp_lty = 1,
  notch = F,
  width = NULL,
  varwidth = F,
  notch.frac = 0.5,
  border = par("fg"),
  staplewex = 0,
  ...
)
}
\arguments{
\item{y}{response variable.}

\item{x}{factor.}

\item{jitter_amount}{value to control the random horizontal spread of data points.}

\item{bp_col}{color of boxplots.}

\item{bp_lty}{line type for boxplots.}

\item{notch}{see ?bxp.}

\item{width}{see ?bxp.}

\item{varwidth}{see ?bxp.}

\item{notch.frac}{see ?bxp.}

\item{border}{see ?bxp.}

\item{staplewex}{see ?bxp.}

\item{...}{further graphical arguments.}
}
\description{
Draws boxplots with individual data points
}
\details{
Returns warnings because argumments are jointly passed to 'boxplot' and 'points'. Will solve later...
}
\examples{
y <- runif(10)
x <- c(rep("a", 5), rep("b", 5))
boxplot_pt(y, x, jitter_amount = 0.1)
}
\keyword{boxplot}
