% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distotable.R
\name{distotable}
\alias{distotable}
\title{Distance Matrix To Table}
\usage{
distotable(dm)
}
\arguments{
\item{dm}{Distance matrix.}
}
\description{
Converts a distance matrix into a data frame.
}
\examples{
dm <- dist(runif(10))
print(dm)
distotable(dm)
}
\keyword{distance}
\keyword{matrix}
