% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_taxonomy.R
\docType{data}
\name{example_taxonomy}
\alias{example_taxonomy}
\title{Example taxonomy}
\format{
## `example_taxonomy`
A data frame with 7 rows and 7:,
\describe{
  \item{kingdom}{kingdom}                              
  \item{phylum}{phylum}                                    
  \item{class}{class}                                     
  \item{order}{order}                                     
  \item{family}{family}                                    
  \item{genus}{genus}
  \item{species}{species}                                                                         
}
}
\usage{
example_taxonomy
}
\description{
Example taxonomy dataset
}
\keyword{datasets}
