% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_string.R
\name{format_string}
\alias{format_string}
\title{Formats Text Strings For Plotting}
\usage{
format_string(
  s,
  i = NULL,
  b = NULL,
  spt = "_",
  sep = " ",
  pattern = NULL,
  add_terms = NULL
)
}
\arguments{
\item{s}{string of text.}

\item{i}{vector indicating the position of words in strings to show in italics.}

\item{b}{vector indicating the position of words in strings to show in bold-face.}

\item{spt}{character indicating the separator between words.}

\item{pattern}{regex pattern to select words to be omitted in formatting.}

\item{add_terms}{vector with words to be omitted from formatting.}
}
\description{
Formats differently words in a set of text strings for plotting.
}
\details{
At the moment, the maximum length os strings admitted is 10 words. Several terms are excluded by default from formatting: "sp", "sp.", "spp", "spp.", "unclassified", "unidentified", "others".
}
\examples{
x <- c("Fusarium_oxysporum_P1304", "Alternaria_sp._P1555", "Fusarium_oxysporum_f.sp._melonis")
format_string(x, i = 1:4, pattern = "P[0-9]", add_terms = "f.sp.")
}
\keyword{format}
\keyword{text}
