% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsChart.R
\name{gainsChart}
\alias{gainsChart}
\title{Generate a gains chart}
\usage{
gainsChart(y, yhat, y2 = NULL, numBins = 10)
}
\arguments{
\item{y}{logical, integer or numeric vector (dependent variable)}

\item{yhat}{numeric vector (predicted values of \code{y})}

\item{y2}{logical, integer or numeric vector (other dependent variable for combined models)}

\item{numBins}{integer value >= 2; number of desired bins}
}
\value{
A data frame with class "\code{mt_gainsChart}" containing the 
following columns:
\itemize{
  \item \code{yhatBin}: integer specifying the bin number (i.e. decile)
  \item \code{minYhat}: minimum value of \code{yhat}
  \item \code{maxYhat}: maximum value of \code{yhat}
  \item \code{Freq}: number of observations (relative)
  \item \code{cumFreq}: number of observations (cumulative)
  \item \code{sumY}: sum of \code{y} values (relative)
  \item \code{sumYB} (combined only): sum of \code{y} (binary) values (relative)
  \item \code{avgY}: average of \code{y} values (relative)
  \item \code{avgYB} (combined only): average of \code{y} (binary) values (relative)
  \item \code{Lift}: lift of \code{avgY} (relative)
  \item \code{LiftYB} (combined only): lift of \code{avgYB} (relative)
  \item \code{sumYC} (combined only): sum of \code{y} (continuous) values (relative)
  \item \code{avgYC} (combined only): average of \code{y} (continuous) values (relative)
  \item \code{LiftYC} (combined only): lift of \code{avgYC} (relative)
  \item \code{perFreqYC} (combined only): average (per Freq) of \code{y} (continuous) values (relative)
  \item \code{LiftYC2} (combined only): lift of \code{perFreqYC} (relative)
  \item \code{cumSumY}: sum of \code{y} values (cumulative)
  \item \code{cumSumYC} (combined only): sum of \code{y} (continuous) values (cumulative)
  \item \code{cumAvgY}: average of \code{y} values (cumulative)
  \item \code{cumLift}: lift of \code{avgY} (cumulative)
  \item \code{cumPctY}: percentage of \code{y} values (cumulative)
  \item \code{cumPctYC} (combined only): percentage of \code{y} (continuous) values (cumulative)
  \item \code{nonY} (binary only): sum of \code{non-y} values (relative)
  \item \code{cumNonY} (binary only): sum of \code{non-y} values (cumulative)
  \item \code{cumPctNonY} (binary only): percentage of \code{non-y} values (cumulative)
  \item \code{cumPctRandom} (continuous & combined): uniform random percentage (cumulative)
  \item \code{KS}: K-S statistic
}
}
\description{
\code{gainsChart} produces a gains chart table, evaluating 
\code{y} as the dependent variable and \code{yhat} as the vector of predicted values. 
When \code{yhat} is a combined model score, use \code{y2} to supply the other 
dependent variable. Users may specify the number of bins (i.e. 10 for deciles) that are desired.
}
\examples{
# pull in sample scored data frame
x <- modelSampleScored
# filter to validation subset
x <- x[x$ValidateFlag, ]
# gains chart binary
gainsChart(x$TargetFlag, x$pTargetFlag)
# gains chart continuous
gainsChart(x[x$TargetFlag, "TargetValue"], x[x$TargetFlag, "pTargetValue"])
# gains chart combined
gainsChart(x$TargetFlag, x$pTargetFlag*x$pTargetValue, x$TargetValue)
}
\seealso{
\code{\link{glm}, \link{lm}}
}
