% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsChartDT.R
\name{gainsChartDT}
\alias{gainsChartDT}
\title{Produce a gains chart datatable}
\usage{
gainsChartDT(x, scoreFormatType = "int", scoreFormatDigits = 5,
  contSumFormatType = "dlr", contSumFormatDigits = 0,
  contAvgFormatType = "dlr", contAvgFormatDigits = 0, KScolor = "#FF3005")
}
\arguments{
\item{x}{\code{mt_gainsChart} object}

\item{scoreFormatType}{character string; format type for score field(s); valid values 
are "int", "dlr" and "pct"}

\item{scoreFormatDigits}{non-negative integer value; number of decimal places 
for score field(s)}

\item{contSumFormatType}{character string; format type for continuous sum field(s); valid values 
are "int", "dlr" and "pct"}

\item{contSumFormatDigits}{non-negative integer value; number of decimal places 
for continuous sum field(s)}

\item{contAvgFormatType}{character string; format type for continuous average (mean) field(s); valid values 
are "int", "dlr" and "pct"}

\item{contAvgFormatDigits}{non-negative integer value; number of decimal places 
for continuous average (mean) field(s)}

\item{KScolor}{character string; text color for max KS value (valid color)}
}
\value{
A \code{datatable} representation of a \code{mt_gainsChart} object.
}
\description{
\code{gainsChartDT} produces a nicely formatted \code{DT::datatable} 
representation of a \code{mt_gainsChart} object.
}
\examples{
# pull in sample scored data frame
x <- modelSampleScored
# filter to validation subset
x <- x[x$ValidateFlag, ]
# gains chart binary
y <- gainsChart(x$TargetFlag, x$pTargetFlag)
gainsChartDT(y)
# gains chart continuous
y <- gainsChart(x[x$TargetFlag, "TargetValue"], x[x$TargetFlag, "pTargetValue"])
gainsChartDT(y)
# gains chart combined
y <- gainsChart(x$TargetFlag, x$pTargetFlag*x$pTargetValue, x$TargetValue)
gainsChartDT(y)
}
\seealso{
\code{\link{gainsChart}, \link{datatable}}
}
