% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsChartPerfCurve.R
\name{gainsChartPerfCurve}
\alias{gainsChartPerfCurve}
\title{Produce a cumulative performance graph based on gains chart(s)}
\usage{
gainsChartPerfCurve(xb, xv = NULL, cb = "#009DDC", cv = "#77BF30",
  cr = "#666666", xLabel = "Cumulative Total Percent",
  yLabel = "Cumulative Outcome Percent")
}
\arguments{
\item{xb}{\code{mt_gainsChart} object}

\item{xv}{\code{mt_gainsChart} object}

\item{cb}{character string; line color for \code{xb} (valid color)}

\item{cv}{character string; line color for \code{xv} (valid color)}

\item{cr}{character string; random/reference line color (valid color)}

\item{xLabel}{character string; x-axis label}

\item{yLabel}{character string; y-axis label}
}
\value{
A \code{\link{plot_ly}} visualization of a 
mt_\code{\link{gainsChart}} object, showing the cumulative percent of total Y 
(\code{cumPctY} for "binary"/"continuous"; \code{cumPctYC} for "combined") by quantile (
\code{yhatBin}) percent.
}
\description{
\code{gainsChartPerfCurve} produces a cumulative performance graph based on a 
mt_\code{\link{gainsChart}} object. When two such objects are supplied, the 
function produces a combined graph, where the first object (\code{xb}) is 
labeled as "Build" and the second object (\code{xv}) is labeled as "Validate".
}
\examples{
# pull in sample scored data frame
x <- modelSampleScored
# filter to validation subset
x <- split(x, x$ValidateFlag)
# gains chart binary
y <- lapply(x, function(z){gainsChart(z$TargetFlag, z$pTargetFlag)})
# cumulative performance graph
gainsChartPerfCurve(y[[1]], y[[2]])
# gains chart continuous
y <- lapply(x, function(z){
    gainsChart(z[z$TargetFlag, "TargetValue"], z[z$TargetFlag, "pTargetValue"])
})
# cumulative performance graph
gainsChartPerfCurve(y[[1]], y[[2]])
# gains chart combined
y <- lapply(x, function(z){
    gainsChart(z$TargetFlag, z$pTargetFlag*z$pTargetValue, z$TargetValue)
})
# cumulative performance graph
gainsChartPerfCurve(y[[1]], y[[2]])
}
\seealso{
\code{\link{gainsChart}, \link{plot_ly}}
}
