% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_coords_x.R
\name{get_node_coords_x}
\alias{get_node_coords_x}
\title{Finds X Coordinates In A Tree Node}
\usage{
get_node_coords_x(tree, node)
}
\arguments{
\item{tree}{tree in 'ape' format.}

\item{node}{number or character, tree node to get coordinate from.}
}
\description{
Finds x coordinates in an 'ape' tree node.
}
\details{
Code taken from https://stackoverflow.com/questions/25624986/draw-on-a-phylogeny-edge-with-r
}
\examples{
get_node_coords_y()
}
\keyword{phylogeny}
