% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tax_info.R
\name{get_tax_info}
\alias{get_tax_info}
\title{Loop to retrieve tanonomic names using myTAI::taxonomy}
\usage{
get_tax_info(
  taxlist,
  ranks = c("genus", "family", "order", "species"),
  db = "ncbi"
)
}
\arguments{
\item{taxlist}{a vector of taxonomic names.}

\item{ranks}{a vector with the taxonomic ranks at which retrieve the names.}

\item{db}{the source database, as specified by myTAI::taxonomy.}
}
\value{
A data frame with the taxonomic names.
}
\description{
Retrieves taxon names at selected taxonomic ranks from a list of provided taxa
}
\keyword{taxonomy}
