% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParms.R
\name{getParms}
\alias{getParms}
\title{Get a list of parameters called in a function}
\usage{
getParms(which = -1)
}
\arguments{
\item{which}{if negative then the number of frames to go back; otherwise the
frame number}
}
\value{
A list of length 1 character vectors containing the parameters that
where supplied to a function. The names of the list are the arguments to
the function.
}
\description{
\code{getParms()} produces a named list of character values
containing the parameters supplied to a function
}
\examples{
foo <- function(x, y){getParms()}
foo(iris, 1983)
}
\seealso{
\code{\link{match.call}, \link{sys.parent}}
}
