% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoValue.R
\name{infoValue}
\alias{infoValue}
\title{Extract the information value from a weight of evidence table}
\usage{
infoValue(x)
}
\arguments{
\item{x}{mt_woe data frame}
}
\value{
A numeric value with class "\code{mt_infoValue}" representing the 
total information value from the input weight of evidence table.
}
\description{
\code{infoValue} returns the total information value from a 
\code{miscTools} weight of evidence table (data frame with the "mt_woe" 
class, i.e. the result of running \code{miscTools::woe()})
}
\examples{
x <- woe(mtcars$hp, mtcars$mpg > 20)
print(x)
class(x)
y <- infoValue(x)
print(y)
class(y)
}
\seealso{
\code{\link{woe}}
}
