% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isString.R
\name{isString}
\alias{isString}
\title{Determine if an object is a string}
\usage{
isString(...)
}
\arguments{
\item{...}{one or more objects to evaluate}
}
\value{
A logical vector indicating whether the input objects are strings.
}
\description{
\code{isString()} determines whether an object is a string. A 
string is defined as a length one character vector.
}
\examples{
isString("abc", 123, c("def", "ghi"))
}
