\name{margEff}
\alias{margEff}
\title{Method for Returning Marginal Effects}

\description{
Currently, this package just defines the generic function \code{margEff}
so that it can be used 
to define \code{margEff} methods for objects of specific classes
in other packages.
}

\usage{
margEff( object, \dots )
}

\arguments{
  \item{ object }{an object of which marginal effects should be calculated.}
  \item{\dots}{further arguments for methods}
}

\author{Arne Henningsen}
\keyword{methods}
