\name{miscTools-internal}
\alias{checkNames}
\alias{logDataSet}
% Document the following:

%%%%


\title{ Undocumented miscTools Functions }

\description{
   Undocumented miscTools Functions
}

\details{
   These are various methods or functions
   waiting for proper documentation to be written :).
}

\keyword{ internal }
