% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEDA.R
\name{modelEDA}
\alias{modelEDA}
\title{Exploratory data analysis on a modeling dataset}
\usage{
modelEDA(x, yname, ytype, bin.numBins = 10, bin.equalBinSize = FALSE,
  bin.minPct = 0, bin.maxPct = 100, nrsq.na.rm = FALSE,
  variable_cluster.n = max(floor((length(x) - 1)/10), 2),
  variable_cluster.na.rm = TRUE, univariateSummary.FUN = mean,
  univariateGraph.yLabel = yname, univariateGraph.yType = ifelse(ytype == 1,
  "pct", "dlr"), univariateGraph.yDigits = ifelse(ytype == 1, 1, 0),
  univariateGraph.yRangeMode = "tozero",
  univariateGraph.barColor = "#BDDFF7",
  univariateGraph.lineColor = "#000000")
}
\arguments{
\item{x}{data frame; a modeling dataset}

\item{yname}{character string; dependent variable column name}

\item{ytype}{integer value of 1 (binary y) or 2 (continuous y)}

\item{bin.numBins}{integer value >= 2; number of desired bins}

\item{bin.equalBinSize}{logical value; return equally sized (TRUE) or equally spaced (FALSE) bins}

\item{bin.minPct}{integer between 0 and 100 specifying a percentile to force as the max endpoint for the low (first) bin}

\item{bin.maxPct}{integer between 0 and 100 specifying a percentile to force as the min endpoint for the high (last) bin (must be > minPct)}

\item{nrsq.na.rm}{logical value indicating whether missing values of x (and their corresponding y values) should be removed}

\item{variable_cluster.n}{integer value >= 2; number of desired clusters}

\item{variable_cluster.na.rm}{logical value; should records with missing values be removed?}

\item{univariateSummary.FUN}{function to be applied to y}

\item{univariateGraph.yLabel}{character string; y variable label}

\item{univariateGraph.yType}{character string; y variable format type; valid values are "int", "dlr" and "pct"}

\item{univariateGraph.yDigits}{non-negative integer value indicating the number of decimal places to show for values of the y variable}

\item{univariateGraph.yRangeMode}{character string; "tozero" (y-axis starts at 0) or "auto" (y-axis extremes determined by data)}

\item{univariateGraph.barColor}{character string; fill color for bars (valid color)}

\item{univariateGraph.lineColor}{character string; line color (valid color)}
}
\value{
A named list with class \code{mt_modelEDA} containing the following 
objects:
\itemize{
  \item \code{dataSummaryDT}: \code{datatable()} object summarizing data
  \item \code{y.relativeHistogram}: \code{relativeHistogram()} object for \code{y}
  \item \code{variable_cluster}: data frame with \code{variable_cluster()} results
  \item \code{woe} (binary \code{y}): named list of \code{woe()} tables
  \item \code{infoValue} (binary \code{y}): named list of \code{infoValue()} values
  \item \code{woeDT} (binary \code{y}): named list of \code{woeDT()} objects
  \item \code{nrsq} (continuous \code{y}): named list of \code{nrsq()} values
  \item \code{variable_cluster_plus}: data frame with \code{variable_cluster_plus()} results
  \item \code{clusterDT}: named list of \code{clusterDT()} objects
  \item \code{univariateSummary}: named list of \code{univariateSummary()} tables
  \item \code{univariateSummaryDT}: named list of \code{univariateSummaryDT()} objects
  \item \code{univariateGraph}: named list of \code{univariateGraph()} objects
}
}
\description{
\code{modelEDA} performs standard exploratory data analysis on a 
modeling dataset, including variable clustering, weight of evidence (binary y), 
numeric R-squared (continuous y) and univariate summaries/graphs. Parameters 
prefaced with a function dot (i.e. \code{bin.}) are only applicable to the use 
of that function.
}
\examples{
# binary y 
x <- modelEDA(mtcars, "vs", 1)
names(x)
x$woeDT

# continuous y
x <- modelEDA(mtcars, "mpg", 2)
names(x)
x$clusterDT
}
\seealso{
\code{\link{woe}, \link{nrsq}, \link{variable_cluster}, \link{univariateGraph}}
}
