% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEDA_Report.R
\name{modelEDA_Report}
\alias{modelEDA_Report}
\title{Produce a standard modelEDA report}
\usage{
modelEDA_Report(x, projectLabel = "Model EDA",
  outFile = "./modelEDA_Report.html", tempDir = ".tempDir")
}
\arguments{
\item{x}{\code{mt_modelEDA} object}

\item{projectLabel}{character string; description header for report home page}

\item{outFile}{character string; file path/name for saving report (*.html)}

\item{tempDir}{character string; temporary directory path}
}
\value{
No object is returned. The report is saved to the specified location.
}
\description{
\code{modelEDA_Report} produces a \code{\link{flexdashboard}} 
report based on a \code{mt_modelEDA} object.
}
\examples{
# binary target
modelEDA_Report(modelEDA(mtcars, "vs", 1))

# continuous target
modelEDA_Report(modelEDA(
    x = mtcars,
    yname = "mpg",
    ytype = 2,
    univariateGraph.yLabel = "Average MPG",
    univariateGraph.yType = "int",
    univariateGraph.yDigits = 1
))
}
\seealso{
\code{\link{modelEDA}, \link{flex_dashboard}}
}
