% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerformance_Report.R
\name{modelPerformance_Report}
\alias{modelPerformance_Report}
\title{Produce a standard model performance report}
\usage{
modelPerformance_Report(x, projectLabel = "Model Performance",
  outFile = "./modelPerformance_Report.html", tempDir = ".tempDir")
}
\arguments{
\item{x}{\code{mt_modelPerformance} object}

\item{projectLabel}{character string; descriptive header for report}

\item{outFile}{character string; file path/name for saving report (*.html)}

\item{tempDir}{character string; temporary directory path}
}
\value{
No object is returned. The report is saved to the specified location.
}
\description{
\code{modelPerformance_Report} produces a \code{\link{flexdashboard}} 
report based on a \code{mt_modelPerformance} object.
}
\examples{
# pull in sample scored data frame
x <- modelSampleScored
str(x)
head(x)
# binary target
# create mt_modelPerformance object
# note that "v" and the "variableImportance..." parms are optional
# in case you do not have a validation set or do not know/care about the variable importance
z <- modelPerformance(
    y = x$TargetFlag,
    yhat = x$pTargetFlag,
    v = x$ValidateFlag
)
modelPerformance_Report(z)
# continuous target
z <- modelPerformance(
    y = x[x$TargetFlag, ]$TargetValue,
    yhat = x[x$TargetFlag, ]$pTargetValue,
    v = x[x$TargetFlag, ]$ValidateFlag,
    gainsChartBarGraph.yType = "dlr",
    gainsChartBarGraph.yDigits = 0
)
modelPerformance_Report(z)
# combined target (with arbitrary variable importance parms)
z <- modelPerformance(
    y = x$TargetFlag,
    yhat = x$pTargetFlag*x$pTargetValue,
    y2 = x$TargetValue,
    v = x$ValidateFlag,
    variableImportanceGraph.x = c(
        "Days Since Last Transaction",
        "Average Order Value",
        "Signed Up Online",
        "Number of Transactions",
        "Tenure",
        "Multi-Category Flag"
    ),
    variableImportanceGraph.y = c(.18, .4, .04, .25, .12, .01)
)
modelPerformance_Report(z)
}
\seealso{
\code{\link{modelPerformance}, \link{flex_dashboard}}
}
