% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsq.R
\name{nrsq}
\alias{nrsq}
\title{Calculate a coefficient of determination (R-squared)}
\usage{
nrsq(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector (independent variable)}

\item{y}{numeric vector (dependent variable)}

\item{na.rm}{logical value indicating whether missing values of \code{x} 
(and their corresponding \code{y} values) should be removed}
}
\value{
A numeric value with class "\code{mt_nrsq}" that indicates the 
proportion of the variance in the dependent variable that is predictable 
from the independent variable.
}
\description{
\code{nrsq()} calculates the coefficient of determination, or 
R-squared value, between an independent variable \code{x} and a dependent 
variable \code{y}. Note that when \code{na.rm = FALSE} (default), missing 
values of \code{x} will be replaced with \code{1} and a flag identifying the 
missings will be included as an additive term in the model. Otherwise, only 
non-missing records are fit.
}
\examples{
# basic example without missing values
nrsq(mtcars$hp, mtcars$mpg)

# add some missing values
x <- ifelse(runif(length(mtcars$hp)) < 0.30, NA, mtcars$hp)
# include missings with adjustment
nrsq(x, mtcars$mpg, na.rm = FALSE)
# exclude missings entirely
nrsq(x, mtcars$mpg, na.rm = TRUE)

# evaluate a whole data frame of predictors
temp <- lapply(mtcars[c("disp", "hp", "wt", "qsec")], nrsq, y = mtcars$mpg)
data.frame(
    VarName = names(temp),
    RSquared = as.numeric(unlist(temp)),
    stringsAsFactors = FALSE
)
}
\seealso{
\code{\link{lm}}, \code{\link{summary.lm}}
}
