% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTemplate.R
\name{parseTemplate}
\alias{parseTemplate}
\title{Parse a template}
\usage{
parseTemplate(fileIn, ...)
}
\arguments{
\item{fileIn}{path to template file (quoted string)}

\item{...}{objects to use as string replacements}
}
\value{
A string representing the contents of \code{fileIn} after the 
specified replacements have been made. As the string is a length one 
character vector, new lines (\code{\\n}) have been inserted accordingly.
}
\description{
\code{parseTemplate()} reads \code{fileIn} and uses 
\code{sprintf} to make string replacements.
}
\examples{
x <- c("My \%s favorite", "\%s is a \%s.")
write(x, "x.txt")
system("more x.txt")
y <- parseTemplate("x.txt", "least", "animal", "cat")
print(y)
write(y, "y.txt")
system("more y.txt")
file.remove("x.txt", "y.txt")
}
\seealso{
\code{\link{sprintf}}
}
