% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_tree.R
\name{plot_group_tree}
\alias{plot_group_tree}
\title{Plot APE Tree With Clade Triangles}
\usage{
plot_group_tree(tree, groups = NULL, n_min = 5, yvar = 1, tree_col = 1, ...)
}
\arguments{
\item{tree}{ape tree object.}

\item{groups}{a character vector with the node labels to collapse}

\item{n_min}{if groups is NULL, all terminal clades containing a minimum of tips are collapsed}

\item{yvar}{value between 1 and 0 to modify the vertical span of triangles}

\item{tree_col}{color for tree edges and triangles}

\item{...}{further arguments for tree (see plot.phylo)}
}
\value{
Plots a phylogenetic tree.
}
\description{
Plots an APE tree tip in which selected clades are collapsed into triangles
}
\examples{
plot_group_tree()
}
\keyword{phylogeny}
