% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Default plotting theme for ggplot2}
\usage{
plot_theme(...)
}
\arguments{
\item{...}{elements from 'theme'}
}
\description{
Defines default graphical parameters for ggplot2 plots.
}
\keyword{plotting}
