% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeHistogram.R
\name{relativeHistogram}
\alias{relativeHistogram}
\title{Produce a relative histogram}
\usage{
relativeHistogram(data, xLabel, barColor = "#BDDFF7")
}
\arguments{
\item{data}{\code{mt_univariateSummary} data frame}

\item{xLabel}{character string; \code{x} variable label}

\item{barColor}{character string; fill color for bars (valid color)}
}
\value{
A \code{\link{plot_ly}} visualization of a 
mt_\code{\link{univariateSummary}} object, showing the distribution of a 
binned variable \code{x}.
}
\description{
\code{relativeHistogram()} produces a relative histogram based on 
the contents of the input data frame, which must be a mt_\code{\link{univariateSummary}} 
object.
}
\examples{
# numeric x with missings
x <- c(rnorm(5000, 100, 10), rep(as.numeric(NA), 500))
z <- univariateSummary(x, numBins = 20)
relativeHistogram(z, "Random Normal")

# logical x
x <- round(runif(20))
z <- univariateSummary(x)
relativeHistogram(z, "Random Binary", "orange")

# character x
relativeHistogram(univariateSummary(iris$Species), "Species", "#77BF30")
}
\seealso{
\code{\link{univariateSummary}}, \code{\link{plot_ly}}
}
