% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_tree_tips.R
\name{sort_tree_tips}
\alias{sort_tree_tips}
\title{Sort APE Tree Tips}
\usage{
sort_tree_tips(tree)
}
\arguments{
\item{tree}{ape tree object.}
}
\value{
A list with the tip labels sorted as in the tree.
}
\description{
Sorts APE tree tip labels after the tree has been modified (e.g. with 'ladderize')
}
\examples{
sort_tree_tips()
}
\keyword{phylogeny}
