% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_vector.R
\name{strsplit_vector}
\alias{strsplit_vector}
\title{Splits A Vector Of Strings}
\usage{
strsplit_vector(x, sep, i)
}
\arguments{
\item{x}{vector with text strings.}

\item{sep}{character separating elements.}

\item{i}{index of elements to retain.}
}
\description{
Splits a vector of strings by a character and returns elements within a position.
}
\examples{
x <- c("Fusarium_oxysporum_P1304", "Alternaria_sp._P1555", "Fusarium_oxysporum_f.sp._melonis")
strsplit_vector(x, sep = "_", i = 1)
}
\keyword{split}
\keyword{text}
