% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateSummary.R
\name{univariateSummary}
\alias{univariateSummary}
\title{Generate a univariate summary table}
\usage{
univariateSummary(x, y = rep(0, length(x)), ytype = 1, FUN = mean, ...)
}
\arguments{
\item{x}{character, factor, logical or numeric vector (independent variable)}

\item{y}{logical, integer or numeric vector (dependent variable)}

\item{ytype}{integer value of 1 (binary y) or 2 (continuous y)}

\item{FUN}{function to be applied to \code{y}}

\item{...}{further arguments passed to \code{\link{bin}}}
}
\value{
A data frame with class "\code{mt_univariateSummary}" containing the 
following columns:
\itemize{
  \item \code{xbin}: binned values of x (character)
  \item \code{Freq}: frequency of observations (integer)
  \item \code{Percent}: relative frequency of observations (numeric)
  \item \code{yagg}: aggregated summary statistic of y (numeric)
}
}
\description{
\code{univariateSummary()} produces a univariate summary table, 
evaluating \code{x} as the independent variable and \code{y} as the dependent 
variable. It uses the S3 generic function \code{\link{bin}} to bin the values of 
\code{x}, then per bin, applies \code{FUN} to \code{y}. The vector \code{x} can include 
missing values but \code{y} cannot. When \code{ytype = 1}, \code{y} must 
contain only unique (coercable) values of 0 or 1. When \code{ytype = 2}, 
\code{y} must have at least 3 distinct values.
}
\examples{
# character x, continuous y
z <- univariateSummary(iris$Species, iris$Sepal.Length, ytype = 2)
print(z)
class(z)

# numeric x, logical y, additional arguments to bin()
z <- univariateSummary(mtcars$hp, mtcars$mpg > 15, numBins = 2)
print(z)
class(z)
}
\seealso{
\code{\link{bin}}
}
