% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateSummaryDT.R
\name{univariateSummaryDT}
\alias{univariateSummaryDT}
\title{Produce a univariate summary datatable}
\usage{
univariateSummaryDT(data, yLabel, yType, yDigits)
}
\arguments{
\item{data}{\code{mt_univariateSummary} data frame}

\item{yLabel}{character string; \code{y} variable label}

\item{yType}{character string; \code{y} variable format type; valid 
values are \code{"int"}, \code{"dlr"} and \code{"pct"}}

\item{yDigits}{non-negative integer value indicating the number of decimal 
places to show for values of the \code{y} variable}
}
\value{
A \code{datatable} representation of a \code{mt_univariateSummary} object.
}
\description{
\code{univariateSummaryDT()} produces a nicely formatted 
\code{DT::datatable} representation of a \code{mt_univariateSummary} object.
}
\examples{
x <- univariateSummary(rnorm(5000), round(runif(5000)))
univariateSummaryDT(x, "Random Binary Rate", "pct", 1)
}
\seealso{
\code{\link{univariateSummary}}, \code{\link{datatable}}
}
