% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_cluster_plus.R
\name{variable_cluster_plus}
\alias{variable_cluster_plus}
\title{Append information value / numeric R-squared to variable clusters}
\usage{
variable_cluster_plus(vc, x)
}
\arguments{
\item{vc}{\code{mt_variable_cluster} object}

\item{x}{list; all elements must inherit either \code{mt_infoValue} or 
\code{mt_nrsq}; element names should be found in the \code{VarName} column in \code{vc}}
}
\value{
A data frame with classes "\code{mt_variable_cluster_plus} and either 
"\code{binary}" or "\code{continuous}" (depending on whether \code{x} contains 
"\code{mt_infoValue}" or "\code{mt_nrsq}" objects, respectively). The returned 
object appends the values and ranks from \code{x} to \code{vc}.
}
\description{
\code{variable_cluster_plus} appends a named list of information 
values or numeric R-squared values to a variable cluster data frame.
}
\examples{
# example with infoValue
vc <- variable_cluster(mtcars, 3)
x <- lapply(mtcars[names(mtcars) != "vs"], woe, y = mtcars$vs)
x <- lapply(x, infoValue)
variable_cluster_plus(vc, x)

# example with nrsq
x <- lapply(mtcars[names(mtcars) != "mpg"], nrsq, y = mtcars$mpg)
variable_cluster_plus(vc, x)
}
\seealso{
\code{\link{variable_cluster}, \link{woe}, \link{infoValue}, 
\link{nrsq}, \link{lapply}}
}
