% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportanceGraph.R
\name{variableImportanceGraph}
\alias{variableImportanceGraph}
\title{Produce a variable importance graph}
\usage{
variableImportanceGraph(x, y, sumYcheck = TRUE,
  barColor = "rgba(0, 55, 82, 0.6)", lineColor = "#003752", lineWidth = 2)
}
\arguments{
\item{x}{character vector}

\item{y}{numeric vector}

\item{sumYcheck}{logical value; check that \code{sum(y) == 1}?}

\item{barColor}{character string; fill color for bars (valid color)}

\item{lineColor}{character string; line color (valid color)}

\item{lineWidth}{non-negative integer value indicating line width (use 0 to omit)}
}
\value{
A horizontal \code{\link{barGraph}}, where \code{x} contains the names 
of the variables in a model and \code{y} contains the model contribution (decimals).
}
\description{
\code{variableImportanceGraph} produces a variable importance graph
}
\examples{
x <- c(
    "Days Since Last Transaction",
    "Average Order Value",
    "Signed Up Online",
    "Number of Transactions",
    "Tenure",
    "Multi-Category Flag"
)
y <- c(.18, .4, .04, .25, .12, .01)
variableImportanceGraph(x, y)
}
\seealso{
\code{\link{barGraph}, \link{plot_ly}}
}
