% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{woe}
\alias{woe}
\title{Generate a weight of evidence (WoE) table}
\usage{
woe(x, y, ...)
}
\arguments{
\item{x}{character, factor, logical or numeric vector (independent variable)}

\item{y}{logical, integer or numeric vector that is coercable to numeric 
with unique values of 0 and 1 (dependent variable)}

\item{...}{further arguments passed to \link{bin}}
}
\value{
A data frame with class "\code{mt_woe}" summarizing the weight of 
evidence and information value calculations for an independent variable 
\code{x} with respect to a binary dependent variable \code{y}.
}
\description{
\code{woe()} produces a weight of evidence table, evaluating 
\code{x} as the independent variable and \code{y} as the dependent variable. 
It depends on an S3 generic function \link{bin} for binning the values of 
\code{x}. Character, factor and logical vectors for \code{x} are all treated 
as categorical. Numeric vectors are treated as continuous (numeric) as long 
as they have more distinct values than the number of bins (\code{numBins}) 
specified, otherwise they are treated as categorical. Integer vectors are 
handled as numeric vectors.
}
\examples{
# factor
woe(iris$Species, round(runif(nrow(iris))))

# numeric - equally spaced vs. equally sized bins
x <- rnorm(1000, mean = 100, sd = 5)
y <- round(runif(1000))
woe(x, y)
woe(x, y, equalBinSize = TRUE)

# evaluate all columns in a data frame
wl <- lapply(mtcars, woe, y = round(runif(nrow(mtcars))))
names(wl)
wl$mpg
# create table of total info values
iv <- unlist(lapply(wl, infoValue))
data.frame(
    Variable = names(iv), 
    InformationValue = as.numeric(iv),
    stringsAsFactors = FALSE
)
}
\seealso{
\code{\link{bin}, \link{infoValue}, \link{woeDT}}
}
