% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woeDT.R
\name{woeDT}
\alias{woeDT}
\title{Produce a weight of evidence (WoE) datatable}
\usage{
woeDT(x)
}
\arguments{
\item{x}{\code{mt_woe} object}
}
\value{
A \code{datatable} representation of a \code{mt_woe} object.
}
\description{
\code{mt_woeDT} produces a nicely formatted \code{DT::datatable} 
representation of a \code{mt_woe} object.
}
\examples{
woeDT(woe(iris$Species, iris$Sepal.Length > 5.75))
}
\seealso{
\code{\link{woe}, \link{datatable}}
}
