% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_tree_tips.R
\name{sort_tree_tips}
\alias{sort_tree_tips}
\title{Sort APE Tree Tips}
\usage{
sort_tree_tips(tree, show_names = T)
}
\arguments{
\item{tree}{ape tree object.}

\item{show_names}{if T, shows tip labels, otherwise shows order.}
}
\value{
A list with the tip labels sorted as in the tree.
}
\description{
Sorts APE tree tip labels after the tree has been modified (e.g. with 'ladderize')
}
\examples{
sort_tree_tips(ape::rtree(10))
}
\keyword{phylogeny}
